+++ src/router/index.js
@@ -26,7 +26,7 @@ const router = createRouter({
     },
     {
       path: '/Microservices',
-      name: Microservices.name,
+      name: 'Microservices',
       component: Microservices,
       icon: 'fas fa-tachometer-alt',
       title: $gettext('Microservices'),
@@ -47,16 +47,17 @@ const router = createRouter({
     },
     {
       path: '/Home',
-      name: Home.name,
+      name: 'Home',
       component: Home,
       icon: 'fa-solid fa-map',
       title: $gettext('Map'),
     },
     {
       path: '/Run',
-      name: Run.name,
+      name: 'Run',
       component: Run,
       icon: 'fa-solid fa-play',
+      loading: false,
       title: $gettext('Parameters and Run'),
       beforeEnter: (to, from, next) => {
         const store = useIndexStore()
@@ -75,7 +76,7 @@ const router = createRouter({
     },
     {
       path: '/ResultMap',
-      name: ResultMap.name,
+      name: 'ResultMap',
       component: ResultMap,
       margin: '5rem',
       icon: 'fa-solid fa-layer-group',
@@ -83,14 +84,14 @@ const router = createRouter({
     },
     {
       path: '/ResultPicture',
-      name: ResultPicture.name,
+      name: 'ResultPicture',
       component: ResultPicture,
       icon: 'fas fa-images',
       title: $gettext('Results Pictures'),
     },
     {
       path: '/ResultTable',
-      name: ResultTable.name,
+      name: 'ResultTable',
       component: ResultTable,
       icon: 'fas fa-table',
       title: $gettext('Result Table'),